src/mlproject/utils/common.py
import os
from box.exceptions import BoxValueError
import yaml
from mlproject import logger 
import json 
import joblib 
from typeguard import typechecked
from box import ConfigBox
from pathlib import Path 
from typing import Any 

@typechecked
def read_yaml(path_to_yaml: Path) -> ConfigBox:
    try:
        with open(path_to_yaml) as yaml_file:
            content = yaml.safe_load(yaml_file)
            if content is None:
                raise ValueError("yaml file is empty")
            logger.info(f"yaml file: {path_to_yaml} loaded successfully.")
            return ConfigBox(content)
    except BoxValueError:
        raise ValueError("yaml file is empty")
    except Exception as e:
        raise e 
    
@typechecked
def create_directories(path_to_directories: list, verbose=True):
    for path in path_to_directories:
        os.makedirs(path, exist_ok=True)
        if verbose:
            logger.info(f"Created directory at: {path}")
            
@typechecked
def save_json(path: Path, data: dict):
    with open(path, "w") as f:
        json.dump(data, f, indent=4)
        
    logger.info(f"json file saved at: {path}")
    
@typechecked
def load_json(path: Path) -> ConfigBox:
    with open(path) as f:
        content = json.load(f)
        
    logger.info(f"json file loaded successfully from:{path}")
    return ConfigBox(content)

@typechecked
def save_bin(data:Any, path: Path):
    joblib.dump(value=data, filename=path)
    logger.info(f"binary file saved at: {path}")
    
@typechecked
def load_bin(path: Path) -> Any:
    data = joblib.load(path)
    logger.info(f"binary file loaded from: {path}")
    return data
 
@typechecked
def get_size(path: Path) -> str:
    size_in_kb = round(os.path.getsize(path)/1024)
    return f"~ {size_in_kb} KB"
           