requirements.txt
pandas
# mlflow==2.2.2
mlflow 
notebook
numpy
scikit-learn
matplotlib
python-box==6.0.2
PyYAML
tqdm
typeguard
joblib
types-PyYAML
types-Flask
Flask-Cors
-e .